split logger function, tensors and backend from main qnn source
